gets rid of unnecessary branch and adds function for containing all markers
